Address thread contention in sending messages
Sometimes more than 1 thread sends messages to the same session. We use
a non-blocking fire-and-forget call to send messages. With this
approach, it should be our responsibility to track that the message got
through. We risk hitting the outgoing buffer limit if we send too
quickly, but this fix is quick and easy.

[DATA-873]

Change-Id: If53390bdce5305b32c4e27ea8cf3a6e4d802577a
Reviewed-on: https://code.brigade.zone/11975
Reviewed-by: Mark Cooper <mark.cooper@brigade.com>
Tested-by: Jim He <jim.he@brigade.com>